Removes RSocketDestination.java
